Rename QuantConnect.API to QuantConnect.Api
Also renames API folders to Api
