+++ Common/Api/AuthenticationResponse.cs
