+++ Common/QuantConnect.csproj
@@ -232,24 +232,24 @@
     <Compile Include="Algorithm\Framework\Alphas\InsightScoreType.cs" />
     <Compile Include="Algorithm\Framework\Portfolio\PortfolioTargetCollection.cs" />
     <Compile Include="AlphaRuntimeStatistics.cs" />
-    <Compile Include="API\AuthenticationResponse.cs" />
-    <Compile Include="API\Backtest.cs" />
-    <Compile Include="API\Compile.cs" />
-    <Compile Include="API\CompileState.cs" />
-    <Compile Include="API\BacktestReport.cs" />
-    <Compile Include="API\Dividends.cs" />
-    <Compile Include="API\Link.cs" />
-    <Compile Include="API\LiveAlgorithm.cs" />
-    <Compile Include="API\LiveAlgorithmResults.cs" />
-    <Compile Include="API\LiveAlgorithmResultsJsonConverter.cs" />
-    <Compile Include="API\LiveAlgorithmSettings.cs" />
-    <Compile Include="API\LiveLog.cs" />
-    <Compile Include="API\Nodes.cs" />
-    <Compile Include="API\Prices.cs" />
-    <Compile Include="API\Project.cs" />
-    <Compile Include="API\ProjectFile.cs" />
-    <Compile Include="API\RestResponse.cs" />
-    <Compile Include="API\Split.cs" />
+    <Compile Include="Api\AuthenticationResponse.cs" />
+    <Compile Include="Api\Backtest.cs" />
+    <Compile Include="Api\Compile.cs" />
+    <Compile Include="Api\CompileState.cs" />
+    <Compile Include="Api\BacktestReport.cs" />
+    <Compile Include="Api\Dividends.cs" />
+    <Compile Include="Api\Link.cs" />
+    <Compile Include="Api\LiveAlgorithm.cs" />
+    <Compile Include="Api\LiveAlgorithmResults.cs" />
+    <Compile Include="Api\LiveAlgorithmResultsJsonConverter.cs" />
+    <Compile Include="Api\LiveAlgorithmSettings.cs" />
+    <Compile Include="Api\LiveLog.cs" />
+    <Compile Include="Api\Nodes.cs" />
+    <Compile Include="Api\Prices.cs" />
+    <Compile Include="Api\Project.cs" />
+    <Compile Include="Api\ProjectFile.cs" />
+    <Compile Include="Api\RestResponse.cs" />
+    <Compile Include="Api\Split.cs" />
     <Compile Include="Benchmarks\FuncBenchmark.cs" />
     <Compile Include="Benchmarks\IBenchmark.cs" />
     <Compile Include="Benchmarks\SecurityBenchmark.cs" />
